Move twilio-related functions to their own folder
