+++ src/index.ts
@@ -8,7 +8,7 @@ import './utils/config';
 import patientRouter from './routes/patient.api';
 import messageRouter from './routes/messages/messages.api';
 import coachRouter from './routes/coach.auth';
-import twilioRouter from './routes/twilio.api';
+import twilioRouter from './routes/twilio/twilio.api';
 import messageTemplateRouter from './routes/messageTemplate.api';
 import RequireHttps from './middleware/require_https';
