+++ src/routes/twilio/twilio.api.ts
@@ -4,18 +4,18 @@ import { ObjectId } from 'mongodb';
 import twilio from 'twilio';
 import bodyParser from 'body-parser';
 
-import { Message } from '../models/message.model';
+import { Message } from '../../models/message.model';
 import {
   TWILIO_ACCOUNT_SID,
   TWILIO_AUTH_TOKEN,
   TWILIO_FROM_NUMBER,
-} from '../utils/config';
+} from '../../utils/config';
 
-import { Outcome } from '../models/outcome.model';
-import { PatientForPhoneNumber } from '../models/patient.model';
-import auth from '../middleware/auth';
-import { parseInboundPatientMessage } from '../domain/message_parsing';
-import { responseForParsedMessage } from '../domain/glucose_reading_responses';
+import { Outcome } from '../../models/outcome.model';
+import { PatientForPhoneNumber } from '../../models/patient.model';
+import auth from '../../middleware/auth';
+import { parseInboundPatientMessage } from '../../domain/message_parsing';
+import { responseForParsedMessage } from '../../domain/glucose_reading_responses';
 
 const twilioClient = twilio(TWILIO_ACCOUNT_SID, TWILIO_AUTH_TOKEN);
 const { MessagingResponse } = twilio.twiml;
